haywire will now keep allocating memory until the request is completely parsed.
Previously, we were assuming that the alloc callback was only called once per request and that the read callback was called when all the data was available.
Unfortunately, that's not the case, and the read callback may be called when only a fraction of the incoming data is available, especially under high concurrency.

To work around this, we allocate/reallocate the request buffer as required to keep up with the incoming data.
Since the buffer may be reallocated to a different memory region, we save the offset for the URL, headers and the body in the original buffer and then repoint all of them when we finish parsing the request, so that they're in the same position in relation to the beginning of the final request buffer.