Adding request size validation. Fixing bug with route comparison.
Now we'll keep an eye on the maximum request size (defaults to 1MB). When that's exceeded, we'll send a 413 error back and close the connection, as there's not point trying to continue reading a request that will exceed the maximum size.
The bug with route comparison had to do with using a stack value instead of the reference and also trying to compare strings using the longest width and not the opposite.
I'm also adding a /ping route that will print the headers and the body, so we can validate them and also get the body back as a response, to ensure that large payloads can be processed correctly.